+++ src/algorithms/marginal_algorithm.cc
@@ -1,13 +1,14 @@
-#include "marginal_algorithm.hpp"
+#include "marginal_algorithm.h"
 
-#include <Eigen/Dense>
 #include <stan/math/prim/fun.hpp>
 
-#include "lib/progressbar/progressbar.hpp"
+#include <Eigen/Dense>
+
+#include "lib/progressbar/progressbar.h"
 #include "marginal_state.pb.h"
-#include "src/collectors/base_collector.hpp"
-#include "src/hierarchies/dependent_hierarchy.hpp"
-#include "src/utils/eigen_utils.hpp"
+#include "src/collectors/base_collector.h"
+#include "src/hierarchies/dependent_hierarchy.h"
+#include "src/utils/eigen_utils.h"
 
 //! \param grid Grid of points in matrix form to evaluate the density on
 //! \param coll Collector containing the algorithm chain
