+++ src/hierarchies/nnig_hierarchy.cc
@@ -1,14 +1,14 @@
-#include "nnig_hierarchy.hpp"
+#include "nnig_hierarchy.h"
 
 #include <google/protobuf/stubs/casts.h>
+#include <stan/math/prim/prob.hpp>
 
 #include <Eigen/Dense>
-#include <stan/math/prim/prob.hpp>
 
 #include "hierarchy_prior.pb.h"
 #include "ls_state.pb.h"
 #include "marginal_state.pb.h"
-#include "src/utils/rng.hpp"
+#include "src/utils/rng.h"
 
 void NNIGHierarchy::initialize() {
   assert(prior != nullptr && "Error: prior was not provided");
