+++ src/hierarchies/nnw_hierarchy.cc
@@ -1,18 +1,18 @@
-#include "nnw_hierarchy.hpp"
+#include "nnw_hierarchy.h"
 
 #include <google/protobuf/stubs/casts.h>
+#include <stan/math/prim/prob.hpp>
 
 #include <Eigen/Dense>
-#include <stan/math/prim/prob.hpp>
 
 #include "hierarchy_prior.pb.h"
 #include "ls_state.pb.h"
 #include "marginal_state.pb.h"
 #include "matrix.pb.h"
-#include "src/utils/distributions.hpp"
-#include "src/utils/eigen_utils.hpp"
-#include "src/utils/proto_utils.hpp"
-#include "src/utils/rng.hpp"
+#include "src/utils/distributions.h"
+#include "src/utils/eigen_utils.h"
+#include "src/utils/proto_utils.h"
+#include "src/utils/rng.h"
 
 //! \param prec_ Value to set to prec
 void NNWHierarchy::set_prec_and_utilities(const Eigen::MatrixXd &prec_) {
