+++ src/includes.h
@@ -1,20 +1,20 @@
 #ifndef BAYESMIX_INCLUDES_H_
 #define BAYESMIX_INCLUDES_H_
 
-#include "algorithms/load_algorithms.hpp"
-#include "algorithms/neal2_algorithm.hpp"
-#include "algorithms/neal8_algorithm.hpp"
-#include "collectors/file_collector.hpp"
-#include "collectors/memory_collector.hpp"
-#include "hierarchies/load_hierarchies.hpp"
-#include "hierarchies/nnig_hierarchy.hpp"
-#include "hierarchies/nnw_hierarchy.hpp"
-#include "mixings/dirichlet_mixing.hpp"
-#include "mixings/load_mixings.hpp"
-#include "mixings/pityor_mixing.hpp"
-#include "runtime/factory.hpp"
-#include "utils/cluster_utils.hpp"
-#include "utils/io_utils.hpp"
-#include "utils/proto_utils.hpp"
+#include "algorithms/load_algorithms.h"
+#include "algorithms/neal2_algorithm.h"
+#include "algorithms/neal8_algorithm.h"
+#include "collectors/file_collector.h"
+#include "collectors/memory_collector.h"
+#include "hierarchies/load_hierarchies.h"
+#include "hierarchies/nnig_hierarchy.h"
+#include "hierarchies/nnw_hierarchy.h"
+#include "mixings/dirichlet_mixing.h"
+#include "mixings/load_mixings.h"
+#include "mixings/pityor_mixing.h"
+#include "runtime/factory.h"
+#include "utils/cluster_utils.h"
+#include "utils/io_utils.h"
+#include "utils/proto_utils.h"
 
 #endif  // BAYESMIX_INCLUDES_H_
