+++ src/mixings/dirichlet_mixing.h
@@ -3,9 +3,9 @@
 
 #include <memory>
 
-#include "base_mixing.hpp"
+#include "base_mixing.h"
 #include "mixing_prior.pb.h"
-#include "src/hierarchies/base_hierarchy.hpp"
+#include "src/hierarchies/base_hierarchy.h"
 
 //! Class that represents the Dirichlet process mixture model.
