+++ src/utils/io_utils.cc
@@ -1,4 +1,4 @@
-#include "io_utils.hpp"
+#include "io_utils.h"
 
 #include <Eigen/Dense>
 #include <fstream>
