+++ src/utils/proto_utils.cc
@@ -1,4 +1,4 @@
-#include "proto_utils.hpp"
+#include "proto_utils.h"
 
 #include <google/protobuf/io/zero_copy_stream_impl.h>
 #include <google/protobuf/text_format.h>
