+++ test/distributions.cc
@@ -1,12 +1,12 @@
-#include "src/utils/distributions.hpp"
+#include "src/utils/distributions.h"
 
 #include <gtest/gtest.h>
+#include <stan/math/prim.hpp>
 
 #include <Eigen/Dense>
-#include <stan/math/prim.hpp>
 #include <vector>
 
-#include "src/utils/rng.hpp"
+#include "src/utils/rng.h"
 
 TEST(mix_dist, 1) {
   auto& rng = bayesmix::Rng::Instance().get();
