+++ test/lpdf.cc
@@ -1,14 +1,14 @@
 #include <gtest/gtest.h>
-
-#include <Eigen/Dense>
 #include <stan/math/prim/fun.hpp>  // lgamma, lmgamma
 #include <stan/math/prim/prob.hpp>
 
+#include <Eigen/Dense>
+
 #include "marginal_state.pb.h"
-#include "src/hierarchies/lin_reg_uni_hierarchy.hpp"
-#include "src/hierarchies/nnig_hierarchy.hpp"
-#include "src/hierarchies/nnw_hierarchy.hpp"
-#include "src/utils/proto_utils.hpp"
+#include "src/hierarchies/lin_reg_uni_hierarchy.h"
+#include "src/hierarchies/nnig_hierarchy.h"
+#include "src/hierarchies/nnw_hierarchy.h"
+#include "src/utils/proto_utils.h"
 
 TEST(lpdf, nnig) {
   NNIGHierarchy hier;
