+++ test/proto_utils.cc
@@ -1,4 +1,4 @@
-#include "src/utils/proto_utils.hpp"
+#include "src/utils/proto_utils.h"
 
 #include <gtest/gtest.h>
