+++ test/write_proto.cc
@@ -2,9 +2,9 @@
 
 #include "ls_state.pb.h"
 #include "marginal_state.pb.h"
-#include "src/hierarchies/nnig_hierarchy.hpp"
-#include "src/hierarchies/nnw_hierarchy.hpp"
-#include "src/utils/proto_utils.hpp"
+#include "src/hierarchies/nnig_hierarchy.h"
+#include "src/hierarchies/nnw_hierarchy.h"
+#include "src/utils/proto_utils.h"
 
 TEST(set_state, uni_ls) {
   double mean = 5;
